backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
psutil>=5.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
apscheduler==3.11.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-storage==3.1.0
google-genai==1.8.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.0
moviepy==2.1.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
passlib==1.7.4
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.4.8
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
pyparsing==3.2.3
pytesseract==0.3.13
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
python-dateutil==2.9.0.post0
gunicorn==23.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
instaloader
sentry-sdk[flask]==2.19.0
flasgger==0.9.7.1
stripe==11.5.0

